Add assertion support for CM kernels
The assertions are implemented with `cm_assert(C)` macro. The macro
takes a condition `C` and if it is false, it prints an error message and
terminates the kernel.

The assertion support is disabled by default, so if the assertion
condition is false, the kernel will not terminate and the error message
will not be printed. To enable the assertion support,
the `CM_ENABLE_ASSERTS` macro should be defined before including the
`cm/cm.h` header.